src/qtcore.js: Implement Repeater-Element.
Implements the QML-element Repeater as well as a class/constructor
JSItemModel allowing to create Models similar to QAbstractItemModel.
The Repeater element currently requires either a integer-value or a
JSItemModel. Other model types are not supported currently.
The delegate-item for now needs to be defined seperately and then
referred to via it's id.