Add support for Components.
Currently this support only comprises component-files, in-file components
are not supported, yet.
This change requires a reorganisation of the scope handling. Now the scope is
not stored as property for each object and fetched from the QMLDocument on each
access, but stored in a new Component-object that every element stores a
reference to, that is dependent of the component and assigned more intellegently.