Add support for ListModel and ListElement types.
I added an optional options argument to QMLItem and QMLBaseObject constructors
that is an object with options passed as properties by name. That allows
passing additional options for finetuning by name,
like "{manualChildHandling: true}" for prevent QMLBaseObject from handling the
children in order to do it manually (needed for Repeater). This is the only
currently supported option. Probably more will come.