Fix resolvePath to return "" if the filename is "".
If e.g. the source of an image is set to "", that means no
image should be loaded. But if the basePath was not "",
resolvePath used to return the basePath (something like
"qml/"). The concequence is an http-error trying to access
the base directory as an image. Now resolvePath resolves ""
always to "", resulting in the browser not trying to access
a not existing image anymore.