Replace evaluatingProperties by evaluatingProperty
Prior evaluatingProperties was a stack used to store the properties that are
currently evaluated. Due to some other changes in the property-system it was
in fact impossible that more than one item was on the stack at the same time.
Thus this commit replaces the stack by a single variable evaluatingProperty
that holds either the property that is currently evaluated or undefined.