Small refacture of the property-system.
This commit replaces the global registry of propertyUpdaters by a registry of
all properties. Now there is only one function updateProperty left, which gets
the index of a property in the global registry and updates that property. The
same change was made for the getter and the setter. This slightly improves
performance and makes the architecture slightly easier.