Add a working signal-slot system.
This commit adds a signal-slot system for internal use and support to define
and use signals in QML. This commit makes 'createFunction' obsolete as it is
completely replaced by signals. Moreover with this commit I removed
QMLTransientValue as it seems not to make sense anylonger. (Why shouldn't the
propertyChanged-signal be emitted during animations? In Qt it is emitted.)