Refactoring of Property-Handling
This commit refactors how properties look like. It adds a new class QMLProperty
and transforms the functions getProperty, setProperty and updateProperty to the
methods get, set and update of that class. The new QMLProperty objects are
stored in a list called $properties in the element object. Dependencies between
properties are now handled through signals.

This commit breaks anchors.fill! (Will be fixed in the next commit)