New Geometry System
This commit establishes a new geometry system (yet again). This time the
geometry system works as I like it:

There are two function updateHGeometry and updateVGeometry that are called
whenever a property relevant to the geometry changes (width, anchors.right,
implicitWidth, etc). This functions checks which geometry-properties are set and
sets all other properties properly. It is able to prioritize properties and
handle conflicts.
From now on element.left is read-only and element.anchors.left is write-only as
it is in Qt. This is useful to differentiate between explicitly set anchors and
automatically set values.

The new system is by far more straightforward, easier to understand and gives
the fundament of implementing layouts.

THIS COMMIT BREAKS A LOT OF OTHER THINGS!
As the fixes are quite huge, I've left them for an own commit.