Enable garbage collecting
Establish a system to delete elements so that the garbage collector can fetch
them. This delete-method is called e.g. when removing elements from a repeater.

This commit fixes the problem that occoured when removing elements from a
repeater. The bug happened because the deleted element recieved property
updates although it wasn't in the list anymore.