Add automatical resizing when using document.body.
With this commit it will be possible to crete a QmlEngine object with no
arguments. If you do that, QmlWeb will use document.body as root element and
will automatically resize the QML root element's parent to always fill the
window. It is useful if you use QmlWeb for the whole webpage and not only for
one part of it. If you set document.body as root element explicitely there will
be no automatic resizing.