Add support for states and transitions.
This change adds the element types "State", "PropertyChanges" and "Transition".
It also adds the properties "states", "transitions" and "state" to Item.
Moreover it makes the properties "targets" and "properties" of PropertyAnimation
work, which is needed for transitions. It also adjusts the parser to understand
the syntax "targets: [rect1, rect2]" correctly. To QMLBinding it adds a way to
set a evauluated binding-function instead of a src, which is needed to be able
to revert properties to their old binding.