Fix position of repeater items in parent.children.
Prior repeaters added items to the end of the list of children of its parent.
For example with positioners this had the consequence that the order of elements
was often wrong. This commit fixes this by adding the elements in front of the
repeater element itself, like it's done in Qt.
In this run, also the methods splice, pop, etc. of QML-Arrays were extended to
emit a changed signal.

This commit also fixes a segmentation fault while removing elements by tidying
up before removing parent, etc.