Fix positioning of text.
Prior text was often positioned one pixel too low because of a wrong line-height
of the div containing the text.
Another problem was, that the implicit width/height of text was wrong if inside
a component, because updateImplicitWidth/Height was not called after inserting
the text to the scene. Both problems are fixed with this commit.
The latter problem also appeared for text inputs and was also fixed there.

As a side-effect during developing the fix, this commit also adds support for
relative line-heights and adds the property lineHeightMode.