Remove rectangle border completely if invisible.
If a rectangle has a border width of 0 or a transparent border set it should be
handled like no border. Prior both properties were translated to css one by one
leading to problems with positioning when the rectangle had a transparent border
with a width != 0 (which is the default for rectangles). Now in this case the
css border-property is set to "none".