Fix bug with ListModel and Repeater.
Inside Repeater delegates it is possible to directly access the properties of
the model like variables. Prior this only worked, when there was at least one
item in the model initially. If the model was empty in the beginning appending
an item was not possible, because the model would not know about which
properties the model has. This commit changes that. Now this information is
gathered when the first time an item is inserted into the model.