Fix applyProperties for components.
Previously applyProperties was run twice for the inner children of a component.
This is infact needed to be able to resolve all references. But it also has to
happen for the component's root element itself and that didn't happen. Now it
also does, enabling references to the root scope from within the root element of
a component.