Set evaluatingProperty also on property update.
This fixes a bug that bindings like "a && b" were not correctly evaulated, if
the initial value of a was like false, because b was never called and so b was
not marked as dependency. Now as soon as a becomes true, b will be called on the
property update and will be marked as dependency as well.

BUG: 325195