Allow property bindings with return statement.
In QML it is possible to write complex property bindings, included in brackets,
that use a return statement to state the value. This is also important for
signal-handlers which are treated as binding definitions by the parser, as well.
So now it is also possible to usea return statement inside a signal-handler.

This implementation is not 100% compatible with Qt's implementation as we can't
accept using a simple expression and using return at the same time ("if(true) 1;
else return 0;" is accepted by QtQML, but not doable in JS). So we only allow a
simple expression WITHOUT brackes OR a codeblock using return, which then NEEDS
to be enclosed in brackets.

BUG: 325250