Make x depend on parent.left instead of parent.x
If parent.parent.x changes, parent is moved. But parent.x doesn't change, so x
isn't updated either, which isn't a problem as it doesn't change as well. But
the consequence is that left, right and horizontalCenter aren't updated as well,
although they change. By making x depend on parent.left, which DOES change, now
everything is updates correctly.
The same change was done for the vertical position.