Add real support for all properties of Animation.
This commit fixes the setting of 'running' to true or false.
Moreover the commit adds real support for loops, alwaysRunToEnd and pause (The
properties have been there before).
Additionally it fixes the corner case when you changes e.g. the target, 'to',
'from' or 'property' while the animation was running. The animation used to
switch over to the new target/property/to/from leaving the old one as it was.
Now the animation is running to an end and with the old target/property/to/from
and for the next run, the updates target/property/to/from values are used.

BUG: 325252