Work around a bug breaking first item of repeaters
Item's $init method has a loop to call $init of all its child elements.
When one of the child elements is a repeater, it will insert items into the
list, causing the first item of the repeater to be skipped.
Now repeaters call the $init method of their first item themselves, ensuring it
will be initialized. This doesn't fix the bug, but at least makes it work again.