Ignore all binding errors when needed.
When initializing a component, it is possible that the outer context is not yet
fully initialized, causing errors in binding evaluation. These errors must be
caught. Prior only reference errors where caught, but there can also happen
TypeErrors and probably others as well. Now all kinds of errors are caught. (If
there is a real error caused by the user, it will be rethrown later)