Refacture context handling.
This commit contains a complete rewrite of the context handling stuff in QmlWeb.
Now the context is a handles as a stack of all component contexts above, to
reflect how QML actually works.

Moreover now the root QML Item won't be a child of the DOM-node the user gives,
but it will take that node as its DOM-node, reducing the number of DOM-nodes
used. Also now the root element won't be able to use "anchors.centerIn: parent"
anymore, as it does not have a parent anymore (it doesn't have one in Qt as
well). Instead now the implicitWidth and implicitHeight of the root element is
set to reflect the size of the root DOM-node resp. the window size.

This commit brings in a bug with e.g. states in conjunction with components.
Look for the next commit for a fix.

BUG: 325247