Fix bug with margins and left+right anchors.
When using anchors.left AND anchors.right together with margins, the margins
would be applied twice. The same happened for anchors.top and anchors.bottom.
This is now fixed.

Moreover I changed the names of the local margin variables in updateVGeometry to
reflect vertical margins (tM and bM instead of lM and rM).