Interpret child properties as property assignment.
With this commit child elements are created in the "correct" way.
Prior element types had a $addChild property, that would get a meta
element as argument and would create the child. This is more effort
most of the time and is not how QtQml works. Now the element types
define a default property and the children are created by assigning
them as value to the default property (thus children creation is now
started from applyProperties(), instead of construct()).

Moreover this commit brings a new, this time (theoretically) rock
solid implementation of property aliases. Now the alias values are
parsed by the parser and passed as objectName and propertyName to
the engine, instead of as bindin..