When assigning an array to a property, copy it.
While property bindings are passed by the parser to the engine as
string, simple constants like 5, "hello", but also [] are passed as
value (so, 5, "hello" or an empty array). This value is then assigned
to the actual instances. For simple values like integers or strings
this happens by value, but for arrays and objects it happens by
reference and thus all instances shared the same array.

This commit adds a check into the setter, copying the array if needed.