Implement FontLoader type.
See http://qt-project.org/doc/qt-5/qml-qtquick-fontloader.html and
http://qt-project.org/doc/qt-4.7/qml-fontloader.html

This currently depends on smnh's MIT-licensed FontLoader.js library to
get notifications when a font is loaded and trigger all events in
time.

Elements that are using that font through 'font.family: fontId.name'
get their dimensions recalculated correctly when the font loads.

If FontLoader.js is not provided the font is still loaded, but
'status' property is set to Error and dimensions recalculation is
triggered after a fixed timeout (3 seconds). In this case the user
gets a notice in console.log.

Timeout (3 seconds) is the maximum time for a font to load. If font
isn't loaded in this time, status is set to Error. Without
FontLoader.js dimensions recalculations for elements are triggered
after that time.

For both cases (with and without FontLoader.js) if the font takes more
than this time to load, dimensions recalculations for elements that
are using this font will not be triggered or will have no effect.

On the other hand, if the font could not be loaded (with FontLoader.js
enabled), the Error status will be set only when this timeout expires.