In $draw exit if mode!=Canvas instead of mode==DOM
In default $draw code, exit if current mode is not
QMLRenderMode.Canvas. This is better than check that mode not equal to
DOM. Because we can have some new modes, for example "zero"-mode when
we do not need rendering at all. This is useful for serverside or
scripting scenarios.