Remove double calling QMLBaseObject constructor.
When creating component instances we used to call the constructor of
QMLBaseObject twice, once with the internal data of the component and
once with the data of the instance. This was especially needed for
property declarations etc. But by now, all this stuff is done outside
of the constructor and almost everything that was done inside the
constructor had a check, whether it's the first time, the constructor
is called. With an exception of two things, one of which was a mistake
and caused bugs. The other one – which is the only thing to be
actually done twice – is altering the context and is now simply being
done explicitly outside.

PR-URL: https://github.com/qmlweb/qmlweb/pull/29
Reviewed-By: Michael Martin Moro <michael@unetresgrossebite.com>
Reviewed-By: Henrik Rudstrøm <hrudstrom@gmail.com>
Reviewed-By: Сковорода Никита Андреевич <chalkerx@gmail.com>
Reviewed-By: Anton Kreuzkamp <akreuzkamp@web.de>