Properly remove animations at overwriting on init.
During initialization it can happen that a property gets assigned a
binding and later, still during init, gets assigned a constant value.
In this case it wouldn't properly remove the binding and when later,
during evaluating bindings phase, the constant value would get
overwritten again, by the binding. This is not intended.

No it's cleaned correctly.