Root element resizing fixes
1. window.onresize function is changed to addEventListener("resize")
call.
2. In case when qml scene is placed in some dom tag (case 2),
we have to call `self.implicitHeight =` and `self.implicitWidth =`
each time the rootElement changes it's geometry
to reposition child elements of qml scene

Also there is a small note. Because in qmlweb implicitHeight= is chained
to updateHGeometry,
which in turn calls height=, which in turn sets element dom style, the
user have to set
rootElement's width/height css styles to !important.
Maybe it should be noted somewhere in the docs. This note is not
regarding to the current patch.

PR-URL: https://github.com/qmlweb/qmlweb/pull/204
Reviewed-By: Сковорода Никита Андреевич <chalkerx@gmail.com>