Split up src/qtcore into src/engine and src/qtbase
This splits up the files in qtcore into qml-engine related stuff and
generic stuff that reimplements QtCore/QtGui functionality.
This implements part of https://github.com/qmlweb/qmlweb/issues/143.

I decided to rename the qtcore directory to qtbase, because it doesn't
only contain QtCore classes, but also QtGui (which in qt5 are both
part of qtbase).

PR-URL: https://github.com/qmlweb/qmlweb/pull/269
Reviewed-By: Сковорода Никита Андреевич <chalkerx@gmail.com>