Revert passing in meta.context to createObject in construct
This was added in 2c7b240

By forcing the context for createObject in "construct" to that of the
calling context, it stops a new context Object from being created with
the Component's context as a prototype.  This will then cause changes
to the newly created Item's context to creep into the passed in
componentContext.