Component: set Item.parent in construct
This mirrors QML behaviour, which doesn't require the user to manually
set the item's parent after calling 'createObject'.

Currently, anywhere 'construct' or 'createObject' is called internally
in qmlweb, the item's 'parent' is manually set afterward, or the newly
created items are part of the default 'data' property, and
'QMLItem.dataChanged' does the setting of the 'parent' property.

The 'PropertyGrid' Render test has been updated to include a mixture of
built-in 'constructor'-based items, and Components loaded from a file,
to ensure that children ordering remains correct.