QmlWeb: split and move out of the engine dir
The namespace file now resides directly in ./src/ dir and is included
first, immediately followed by the complete qtcore (as it doesn't
depend on anything but the namespace).

Also move the map of keybord events to keyboard.js

The only code change is that now `keyboardSignals` is being exported
as `QmlWeb.keyboardSignals`, everything else is just code movement.

Refs: https://github.com/qmlweb/qmlweb/issues/143
PR-URL: https://github.com/qmlweb/qmlweb/pull/309
Reviewed-By: Anton Kreuzkamp <akreuzkamp@web.de>