tests: add BorderImage render test (partially disabled)
This test is partially disabled (covered with black rectangles), because
PhantomJS currently doesn't support pixelated rendering for border-image

Once that is fixed, this test could be re-rendered without those
black rectanges.

This test actually works on QmlWeb side when viewed e.g. with Chrome, so
the only problem here is on the PhantomJS side.

Even though the most part of the test seem disabled, it actually
currently tests for the correct placement of all four corner images,
which greately reduces the probability of a failure that would get
unnoticed.