engine: Added QMLUrl type
This resolves url-type properties when they are set. This ensures that
they are relative to the file path of the Component they are set within.

The $basePath is now a fully-resolved URL (e.g. http://...). This ensures
that resolved paths are not incorrectly re-resolved by any subsequent
calls to $resolvePath.

Removed calls to $resolvePath from Components that have url properties,
since this is handled automatically.

$resolvePath will now remove duplicate slashes and resolve dot segments.

PR-URL: https://github.com/qmlweb/qmlweb/pull/320
Reviewed-By: Сковорода Никита Андреевич <chalkerx@gmail.com>