Only load from QmlWeb.qrc if URL is qrc://
This removes places that were using QmlWeb.qrc as a cache, as that is not
its purpose.

Added QMLEngine.$resolveImageURL for use by Image and BorderImage to
take an already-resolved source URL and get a DOM-usable URL out of it.
Mainly, in the case of a "qrc://" URL, it will look up a "correct" URL
in QmlWeb.qrc (or return the path component of the URL for
backwards-compatibility).

PR-URL: https://github.com/qmlweb/qmlweb/pull/320