MouseArea: Emit positionChanged when pressed & cursor moves outside
The "mousemove" DOM event is only sent to the MouseArea's DOM element while
the cursor remains over the DOM element. This now attaches a temporary
event listener to "document" and translates the JavaScript MouseEvent
"clientX" and "clientY" properties to the MouseArea's coordinate space
(In regards to CSS transformations, this only takes into account translate
and scale transformations).

PR-URL: https://github.com/qmlweb/qmlweb/pull/356
Reviewed-By: Сковорода Никита Андреевич <chalkerx@gmail.com>