CustomElements: initial Custom Elements v1 support
One could now define QML Custom elements with the following call:

  QmlWeb.registerElement('qml-main', 'qml/main.qml');

And use them as:

  <qml-main width="300" height="300" color="red"></qml-main>

Given that the browser supports Custom Elements v1 or that a polyfill is
loaded.

Refs: https://www.w3.org/TR/custom-elements/

Example included.

Caveat: atm, QmlWeb supports only one running engine, which limits
the number of active elements. That has to be fixed later.

PR-URL: https://github.com/qmlweb/qmlweb/pull/363