modules: register attached objects on their declaration
This is not ideal (attached objects are always registered), but is
still better than what we had earlier — this allows to use attached
objects without manually creating their instances.

This is needed for QtQuick.Window Screen type.

It is not worse than what we have before because those are anyway
being registered on the common QMLBaseObject.prototype for now.