Timer: make parent not read-only for now
Some parts of QmlWeb set .parent property directly, that was causing a
conflict and an exception was being thrown.

Long-term, we should create a better fix, but this fixes the immediate
problem for now, and properties not being read-only when they should
causes much less damage than unexpected exceptions on valid qml code.